fix: correct int type

+++ packages/contracts-bridge/contracts/mocks/MockAccountant.sol
@@ -14,7 +14,6 @@ contract MockNftAccountant is NFTRecoveryAccountant {
         } else {
             totalAffected[_asset] = totalAffected[_asset] + 10 ether;
         }
-
     }
 
     function exposed_record(
@@ -35,7 +34,7 @@ contract MockNftAccountant is NFTRecoveryAccountant {
         require(_user == msg.sender, "only NFT holder can recover");
         Record memory _rec = records[_id];
         uint256 _amount = 50000000;
-        _rec.recovered += _amount;
+        _rec.recovered += uint96(_amount);
         emit Recovery(_id, _rec.asset, _user, _amount);
     }
